+++ jobject-copy/src/main/java/de/alpharogroup/copy/object/CopyObjectQuietlyExtensions.java
@@ -21,9 +21,12 @@
 package de.alpharogroup.copy.object;
 
 import java.io.IOException;
-import java.io.OutputStream;
+import java.io.Serializable;
 import java.lang.reflect.InvocationTargetException;
 
+import org.apache.commons.beanutils.BeanUtils;
+
+import de.alpharogroup.reflection.ReflectionExtensions;
 import lombok.experimental.UtilityClass;
 import lombok.extern.slf4j.Slf4j;
 
@@ -35,6 +38,48 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public final class CopyObjectQuietlyExtensions
 {
+
+	/**
+	 * Copy the given original object to the given destination object. This also works on private
+	 * fields.
+	 *
+	 * @param <ORIGINAL>
+	 *            the generic type of the original object.
+	 * @param <DESTINATION>
+	 *            the generic type of the destination object.
+	 * @param original
+	 *            the original object.
+	 * @param destination
+	 *            the destination object.
+	 * @param fieldName
+	 *            the field name
+	 * @return the destination object
+	 */
+	public static final <ORIGINAL, DESTINATION> DESTINATION copyPropertyWithReflectionQuietly(
+		final ORIGINAL original, final DESTINATION destination, final String fieldName)
+	{
+		try
+		{
+			ReflectionExtensions.copyFieldValue(original, destination, fieldName);
+		}
+		catch (NoSuchFieldException e)
+		{
+			log.error(e.getLocalizedMessage(), e);
+		}
+		catch (SecurityException e)
+		{
+			log.error(e.getLocalizedMessage(), e);
+		}
+		catch (IllegalArgumentException e)
+		{
+			log.error(e.getLocalizedMessage(), e);
+		}
+		catch (IllegalAccessException e)
+		{
+			log.error(e.getLocalizedMessage(), e);
+		}
+		return destination;
+	}
 
 	/**
 	 * Copy quietly the given original object to the given destination object.
@@ -59,18 +104,16 @@ public final class CopyObjectQuietlyExtensions
 		catch (final IllegalAccessException e)
 		{
 			log.error(e.getLocalizedMessage(), e);
-			return null;
 		}
 		catch (final InvocationTargetException e)
 		{
 			log.error(e.getLocalizedMessage(), e);
-			return null;
 		}
 		catch (final IllegalArgumentException e)
 		{
 			log.error(e.getLocalizedMessage(), e);
-			return null;
 		}
+		return null;
 	}
 
 	/**
@@ -125,46 +168,57 @@ public final class CopyObjectQuietlyExtensions
 	}
 
 	/**
-	 * Closes the given OutputStream.
+	 * Copy the given object and return a copy of it. Note: this method decorates the method of
+	 * {@link BeanUtils#copyProperties(Object, Object)} and create a new object for the returned
+	 * object.
 	 *
-	 * @param out
-	 *            The OutputStream to close.
-	 * @deprecated use instead the try-with-resources Statement. Note: will be removed on next minor
-	 *             version
-	 * @return Returns true if the OutputStream is closed otherwise false.
+	 * @param <T>
+	 *            the generic type of the given object.
+	 * @param original
+	 *            the original object.
+	 * @return the new object that is a copy of the given object.
 	 */
-	public static boolean closeOutputStream(OutputStream out)
+	public static final <T> T copyPropertiesQuietly(final T original)
 	{
-		boolean closed = true;
+		T destination = null;
 		try
 		{
-			if (out != null)
-			{
-				out.flush();
-				out.close();
-				out = null;
-			}
+			destination = CopyObjectExtensions.copyProperties(original);
 		}
-		catch (final IOException e)
+		catch (IllegalAccessException | InvocationTargetException | IllegalArgumentException
+			| InstantiationException e)
 		{
-			closed = false;
+			log.error(e.getLocalizedMessage(), e);
 		}
-		finally
+		return destination;
+	}
+
+	/**
+	 * Copy quietly the given Object and returns the copy from the object or null if the object
+	 * can't be serialized.
+	 *
+	 * @param <T>
+	 *            the generic type of the given object
+	 * @param orig
+	 *            The object to copy.
+	 * @return Returns a copy from the original object or null if the object can't be serialized.
+	 */
+	public static <T extends Serializable> T copySerializedObjectQuietly(final T orig)
+	{
+		try
 		{
-			try
-			{
-				if (out != null)
-				{
-					out.flush();
-					out.close();
-				}
-			}
-			catch (final IOException e)
-			{
-				closed = false;
-			}
+			T serializedObject = CopyObjectExtensions.copySerializedObject(orig);
+			return serializedObject;
 		}
-		return closed;
+		catch (ClassNotFoundException e)
+		{
+			log.error(e.getLocalizedMessage(), e);
+		}
+		catch (IOException e)
+		{
+			log.error(e.getLocalizedMessage(), e);
+		}
+		return null;
 	}
 
 }